Update patch set 3
Patch Set 3:

I wonder if we should update the homepage to also have a nice looking landing page?

Similar to https://www.phacility.com/ which makes you go wow then install it?

Obviously it would have to be different but would use the material skin (sorry if i posted this in the wrong place).

I feel like if gerrit had a nice landing page with lots of colour and a image on the front it will make people go wow.

Patch-set: 3
CC: Gerrit User 1029953 <1029953@173816e5-2b9a-37c3-8a2e-48639d4f1153>