Redirect old *.md links to *.html in Firebase config
Since we now serve rendered HTML, the file extension has to change.

In addition, output files now use a flat directory structure; the
filenames in this change correspond to the "permalink" values at the top
of each Jekyll-ized.md file. Our tech writers recommend this flat
directory layout as it means we can rearrange documents in a
hierarchical site outline without having to maintain the hierarchy in
the file tree, avoiding the need to physically move the files.